+++ ndexindraloader/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Indra Annotator"""
 __email__ = 'contact@ndexbio.org'
-__version__ = '0.2.0'
+__version__ = '0.2.0a1'